+++ src/frontend/app/services/specific-angular-markdown.service.ts
@@ -1,0 +1,10 @@
+import {Injectable} from '@angular/core';
+import {ExtendedAngularMarkdownService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/extended-angular-markdown.service';
+import {AngularHtmlService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/angular-html.service';
+
+@Injectable()
+export class SpecificAngularMarkdownService extends ExtendedAngularMarkdownService {
+    constructor(angularHtmlService: AngularHtmlService) {
+        super(angularHtmlService);
+    }
+}