+++ src/frontend/app/tdoc-admin/components/tdoc-createpage/tdoc-createpage.component.spec.ts
@@ -13,7 +13,6 @@ import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generi
 import {ErrorResolver} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/resolver/error.resolver';
 import {PageUtils} from '@dps/mycms-frontend-commons/dist/angular-commons/services/page.utils';
 import {AngularMarkdownService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/angular-markdown.service';
-import {NgxMdModule, NgxMdService} from 'ngx-md';
 import {AngularHtmlService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/angular-html.service';
 import {CommonRoutingService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/common-routing.service';
 import {RouterStub} from '@dps/mycms-frontend-commons/dist/angular-commons/testing/router-stubs';
@@ -30,6 +29,8 @@ import {TourDocRoutingService} from '../../../../shared/tdoc-commons/services/td
 import {TourDocContentUtils} from '../../../shared-tdoc/services/tdoc-contentutils.service';
 import {LayoutService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/layout.service';
 import {ToastrServiceStub} from '@dps/mycms-frontend-commons/dist/testing/toasts-stubs';
+import {SimpleAngularHtmlService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/simple-angular-html.service';
+import {SimpleAngularMarkdownService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/simple-angular-markdown.service';
 
 describe('TourDocCreatepageComponent', () => {
     let component: TourDocCreatepageComponent;
@@ -40,7 +41,6 @@ describe('TourDocCreatepageComponent', () => {
             declarations: [TourDocCreatepageComponent, TourDocDateFormatPipe],
             imports: [
                 TranslateModule.forRoot(),
-                NgxMdModule.forRoot(),
                 NoopAnimationsModule
             ],
             providers: [
@@ -56,9 +56,8 @@ describe('TourDocCreatepageComponent', () => {
                 TourDocRoutingService,
                 { provide: ToastrService, useValue: new ToastrServiceStub() },
                 TranslateService,
-                NgxMdService,
-                AngularMarkdownService,
-                AngularHtmlService,
+                { provide: AngularMarkdownService, useClass: SimpleAngularMarkdownService },
+                { provide: AngularHtmlService, useClass: SimpleAngularHtmlService },
                 ErrorResolver,
                 PageUtils,
                 GenericTrackingService,